pyproject.toml
[project]
name = "changelog-monitor"
version = "0.1.0"
description = "GitHub CHANGELOG monitoring, translation, and LINE notification system"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-genai>=0.2.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
